WIP: Split code for asynchronous execution

+++ server/nodegroup.go
@@ -151,24 +151,29 @@ func (g *AutoScalerServerNodeGroup) AllNodes() []*AutoScalerServerNode {
 	return append(utils.Values(g.Nodes), utils.Values(g.pendingNodes)...)
 }
 
-func (g *AutoScalerServerNodeGroup) setNodeGroupSize(c types.ClientGenerator, newSize int) error {
+func (g *AutoScalerServerNodeGroup) setNodeGroupSize(c types.ClientGenerator, newSize int, prepareOnly bool) ([]*AutoScalerServerNode, error) {
 	glog.Debugf("AutoScalerServerNodeGroup::setNodeGroupSize, nodeGroupID:%s", g.NodeGroupIdentifier)
 
-	var err error
-
 	g.Lock()
+	defer g.Unlock()
 
 	delta := newSize - g.targetSize()
 
 	if delta < 0 {
-		err = g.deleteNodes(c, delta)
+		if prepareOnly {
+			return g.prepareDeleteNodes(delta), nil
+		} else {
+			return g.deleteNodes(c, delta)
+		}
 	} else if delta > 0 {
-		_, err = g.addNodes(c, delta)
+		if prepareOnly {
+			return g.prepareNodes(c, delta)
+		} else {
+			return g.addNodes(c, delta)
+		}
 	}
 
-	g.Unlock()
-
-	return err
+	return []*AutoScalerServerNode{}, nil
 }
 
 func (g *AutoScalerServerNodeGroup) refresh() {
@@ -196,12 +201,7 @@ func (g *AutoScalerServerNodeGroup) findNamedNode(nodeName string) *AutoScalerSe
 	return node
 }
 
-// delta must be negative!!!!
-func (g *AutoScalerServerNodeGroup) deleteNodes(c types.ClientGenerator, delta int) error {
-	glog.Debugf("AutoScalerServerNodeGroup::deleteNodes, nodeGroupID:%s", g.NodeGroupIdentifier)
-
-	var err error
-
+func (g *AutoScalerServerNodeGroup) prepareDeleteNodes(delta int) []*AutoScalerServerNode {
 	pendingNodes := utils.Values(g.pendingNodes)
 	startIndex := len(pendingNodes) - 1
 	tempNodes := make([]*AutoScalerServerNode, 0, -delta)
@@ -213,10 +213,6 @@ func (g *AutoScalerServerNodeGroup) deleteNodes(c types.ClientGenerator, delta i
 		if node.NodeType == AutoScalerServerNodeAutoscaled {
 			delta++
 			tempNodes = append(tempNodes, node)
-
-			if err = node.deleteVM(c); err != nil {
-				glog.Errorf(constantes.ErrUnableToDeleteVM, node.InstanceName, err)
-			}
 		}
 
 		if delta == 0 {
@@ -224,18 +220,41 @@ func (g *AutoScalerServerNodeGroup) deleteNodes(c types.ClientGenerator, delta i
 		}
 	}
 
-	for _, node := range tempNodes {
-		g.RunningNodes[node.NodeIndex] = ServerNodeStateDeleted
-		g.removeNamedNode(node.InstanceName)
+	return tempNodes
+}
+
+func (g *AutoScalerServerNodeGroup) destroyNode(c types.ClientGenerator, node *AutoScalerServerNode) error {
+	g.RunningNodes[node.NodeIndex] = ServerNodeStateDeleted
+	g.removeNamedNode(node.InstanceName)
+
+	return node.deleteVM(c)
+}
+
+func (g *AutoScalerServerNodeGroup) destroyNodes(c types.ClientGenerator, nodes []*AutoScalerServerNode) ([]*AutoScalerServerNode, error) {
+	var err error
+	deletedNodes := make([]*AutoScalerServerNode, 0, len(nodes))
+
+	for _, node := range nodes {
+		if err = g.destroyNode(c, node); err != nil {
+			glog.Errorf(constantes.ErrUnableToDeleteVM, node.InstanceName, err)
+		} else {
+			deletedNodes = append(deletedNodes, node)
+		}
 	}
 
-	return err
+	return deletedNodes, err
+}
+
+// delta must be negative and will delete nodes in pending!!!!
+func (g *AutoScalerServerNodeGroup) deleteNodes(c types.ClientGenerator, delta int) ([]*AutoScalerServerNode, error) {
+	glog.Debugf("AutoScalerServerNodeGroup::deleteNodes, nodeGroupID:%s", g.NodeGroupIdentifier)
+
+	return g.destroyNodes(c, g.prepareDeleteNodes(-delta))
 }
 
 func (g *AutoScalerServerNodeGroup) addManagedNode(crd *v1alpha1.ManagedNode) (*AutoScalerServerNode, error) {
 	controlPlane := crd.Spec.ControlPlane
-	nodeIndex := g.findNextNodeIndex(true)
-	nodeName := g.nodeName(nodeIndex, controlPlane, true)
+	nodeName, nodeIndex := g.nodeName(g.findNextNodeIndex(true), controlPlane, true)
 
 	if awsConfig := g.configuration.GetAwsConfiguration(g.NodeGroupIdentifier); awsConfig != nil {
 		var desiredENI *aws.UserDefinedNetworkInterface
@@ -302,8 +321,7 @@ func (g *AutoScalerServerNodeGroup) addManagedNode(crd *v1alpha1.ManagedNode) (*
 	}
 }
 
-func (g *AutoScalerServerNodeGroup) addNodes(c types.ClientGenerator, delta int) ([]*AutoScalerServerNode, error) {
-	glog.Debugf("AutoScalerServerNodeGroup::addNodes, nodeGroupID:%s", g.NodeGroupIdentifier)
+func (g *AutoScalerServerNodeGroup) prepareNodes(c types.ClientGenerator, delta int) ([]*AutoScalerServerNode, error) {
 
 	tempNodes := make([]*AutoScalerServerNode, 0, delta)
 
@@ -312,10 +330,8 @@ func (g *AutoScalerServerNodeGroup) addNodes(c types.ClientGenerator, delta int)
 		return []*AutoScalerServerNode{}, fmt.Errorf(constantes.ErrNodeGroupNotFound, g.NodeGroupIdentifier)
 	}
 
-	for index := 0; index < delta; index++ {
-
-		nodeIndex := g.findNextNodeIndex(false)
-		nodeName := g.nodeName(nodeIndex, false, false)
+	for {
+		nodeName, nodeIndex := g.nodeName(g.findNextNodeIndex(false), false, false)
 
 		if awsConfig := g.configuration.GetAwsConfiguration(g.NodeGroupIdentifier); awsConfig != nil {
 
@@ -347,13 +363,29 @@ func (g *AutoScalerServerNodeGroup) addNodes(c types.ClientGenerator, delta int)
 			tempNodes = append(tempNodes, node)
 
 			g.pendingNodes[node.InstanceName] = node
+
+			delta--
+
+			if delta == 0 {
+				break
+			}
 		} else {
 			g.pendingNodes = make(map[string]*AutoScalerServerNode)
 			return []*AutoScalerServerNode{}, fmt.Errorf("unable to find node group named %s", g.NodeGroupIdentifier)
 		}
 	}
 
-	return g.createNodes(c, tempNodes)
+	return tempNodes, nil
+}
+
+func (g *AutoScalerServerNodeGroup) addNodes(c types.ClientGenerator, delta int) ([]*AutoScalerServerNode, error) {
+	glog.Debugf("AutoScalerServerNodeGroup::addNodes, nodeGroupID:%s", g.NodeGroupIdentifier)
+
+	if tempNodes, err := g.prepareNodes(c, delta); err == nil {
+		return g.createNodes(c, tempNodes)
+	} else {
+		return tempNodes, err
+	}
 }
 
 // return the list of successfuly created nodes
@@ -377,7 +409,7 @@ func (g *AutoScalerServerNodeGroup) createNodes(c types.ClientGenerator, nodes [
 		}
 
 		if err = node.launchVM(c, g.NodeLabels, g.SystemLabels); err != nil {
-			glog.Errorf(constantes.ErrUnableToLaunchVM, node.NodeName, err)
+			glog.Errorf(constantes.ErrUnableToLaunchVM, node.InstanceName, err)
 
 			node.cleanOnLaunchError(c, err)
 
@@ -725,11 +757,11 @@ func (g *AutoScalerServerNodeGroup) deleteNode(c types.ClientGenerator, node *Au
 	var err error
 
 	if err = node.deleteVM(c); err != nil {
-		glog.Errorf(constantes.ErrUnableToDeleteVM, node.NodeName, err)
+		glog.Errorf(constantes.ErrUnableToDeleteVM, node.InstanceName, err)
 	}
 
 	g.RunningNodes[node.NodeIndex] = ServerNodeStateDeleted
-	g.removeNamedNode(node.NodeName)
+	g.removeNamedNode(node.InstanceName)
 
 	if node.NodeType == AutoScalerServerNodeAutoscaled {
 		g.numOfProvisionnedNodes--
@@ -791,8 +823,9 @@ func (g *AutoScalerServerNodeGroup) getManagedNodePrefix() string {
 	return g.ManagedNodeNamePrefix
 }
 
-func (g *AutoScalerServerNodeGroup) nodeName(vmIndex int, controlplane, managed bool) string {
+func (g *AutoScalerServerNodeGroup) nodeName(vmIndex int, controlplane, managed bool) (string, int) {
 	var start int
+	config := g.configuration.GetAwsConfiguration(g.NodeGroupIdentifier)
 
 	if controlplane {
 		start = 2
@@ -812,17 +845,23 @@ func (g *AutoScalerServerNodeGroup) nodeName(vmIndex int, controlplane, managed 
 		}
 
 		if found := g.findNamedNode(nodeName); found == nil {
-			return nodeName
+			if !config.Exists(nodeName) {
+				return nodeName, vmIndex
+			} else {
+				glog.Warnf(constantes.ErrVMAlreadyExists, nodeName)
+				g.RunningNodes[vmIndex] = ServerNodeStateRunning
+				vmIndex++
+			}
 		}
 	}
 
 	// Should never reach this code
 	if controlplane {
-		return fmt.Sprintf("%s-%s-%02d", g.NodeGroupIdentifier, g.getControlPlanePrefix(), vmIndex-g.numOfExternalNodes-g.numOfProvisionnedNodes-g.numOfManagedNodes+g.numOfControlPlanes+1)
+		return fmt.Sprintf("%s-%s-%02d", g.NodeGroupIdentifier, g.getControlPlanePrefix(), vmIndex-g.numOfExternalNodes-g.numOfProvisionnedNodes-g.numOfManagedNodes+g.numOfControlPlanes+1), vmIndex
 	} else if managed {
-		return fmt.Sprintf("%s-%s-%02d", g.NodeGroupIdentifier, g.getManagedNodePrefix(), vmIndex-g.numOfExternalNodes-g.numOfProvisionnedNodes+1)
+		return fmt.Sprintf("%s-%s-%02d", g.NodeGroupIdentifier, g.getManagedNodePrefix(), vmIndex-g.numOfExternalNodes-g.numOfProvisionnedNodes+1), vmIndex
 	} else {
-		return fmt.Sprintf("%s-%s-%02d", g.NodeGroupIdentifier, g.getProvisionnedNodePrefix(), vmIndex-g.numOfExternalNodes-g.numOfManagedNodes+1)
+		return fmt.Sprintf("%s-%s-%02d", g.NodeGroupIdentifier, g.getProvisionnedNodePrefix(), vmIndex-g.numOfExternalNodes-g.numOfManagedNodes+1), vmIndex
 	}
 }
